+++ scripts/process_qrc.py
@@ -37,9 +37,9 @@ from watchdog.events import FileSystemEventHandler
 from watchdog.observers import Observer
 
 # Local imports
-from qdarkstyle import PACKAGE_PATH
-from qdarkstyle.utils.images import create_images, create_palette_image, generate_qrc_file
-from qdarkstyle.utils.scss import create_qss
+from jam_darkstyle import PACKAGE_PATH
+from jam_darkstyle.utils.images import create_images, create_palette_image, generate_qrc_file
+from jam_darkstyle.utils.scss import create_qss
 
 
 class QSSFileHandler(FileSystemEventHandler):